+++ presentation/src/main/java/com/madrid/presentation/screens/libraryScreen/ViewAllScreen.kt
@@ -26,7 +26,7 @@ import com.madrid.designSystem.theme.Theme
 import com.madrid.presentation.component.SwipeToDeleteCard
 import com.madrid.presentation.navigation.Destinations
 import com.madrid.presentation.navigation.LocalNavController
-import com.madrid.presentation.screens.addtolist.SuccessNotificationRow
+import com.madrid.presentation.component.addtolist.SuccessNotificationRow
 import com.madrid.presentation.screens.libraryScreen.component.LoadingContent
 import com.madrid.presentation.viewModel.libraryViewModel.viewAll.ViewAllEffect
 import com.madrid.presentation.viewModel.libraryViewModel.viewAll.ViewAllInteractionListener
